+++ inst/shiny/modules/tab_data/data_mapping.R
@@ -13,7 +13,10 @@ NON_STD_MAPPING_INFO <- data.frame(
   ),
   mapping_section = c("Supplemental Variables", "Sample Variables"),
   mapping_alternatives = c(
-    "TRTA, TRTAN, ACTARM, TRT01A, TRT01P, RACE, SEX, GROUP, STRAIN, DOSFRM, NOMDOSE, DOSEP",
+    paste0(
+      "TRTA, TRTAN, ACTARM, TRT01A, TRT01P, RACE, SEX, GROUP, ",
+      "STRAIN, DOSFRM, NOMDOSE, DOSEP, COHORT, PART, PERIOD"
+    ),
     ""
   ),
   is_multiple_choice = c(TRUE, TRUE),