+++ inst/shiny/modules/tab_explore.R
@@ -7,16 +7,24 @@ tab_explore_ui <- function(id) {
 
   navset_card_pill(
     id = "visuals",
-    nav_panel("Individual Plots",
+    height = "90vh",
+    full_screen = TRUE,
+    nav_panel(
+      "Individual Plots",
       layout_sidebar(
         sidebar = plot_sidebar_ui(ns("individual_sidebar"), is_mean_plot = FALSE),
-        plotlyOutput(ns("individualplot"), height = "100%")
+        fillable = TRUE,
+        plotlyOutput(ns("individualplot"), height = "100%"),
+        br()
       )
     ),
-    nav_panel("Mean Plots",
+    nav_panel(
+      "Mean Plots",
       layout_sidebar(
         sidebar = plot_sidebar_ui(ns("mean_sidebar"), is_mean_plot = TRUE),
-        plotlyOutput(ns("mean_plot"), height = "100%")
+        fillable = TRUE,
+        plotlyOutput(ns("mean_plot"), height = "100%"),
+        br()
       )
     ),
     nav_panel(